Move bmm dtype validation into meta
